Create html2pdf.js
This is a JavaScript function that defines a click event for an HTML element with an ID of "impress". When this element is clicked, it generates a PDF file using the html2pdf library, which converts an HTML element to a PDF document.

The PDF file generated has the following properties:

The margins are set to 5.
The filename is set to "file_" followed by the current timestamp in milliseconds, and the file extension is ".pdf".
The image type is set to JPEG with a quality of 0.98.
The html2canvas configuration object is used to set options for the html2canvas library, which is used by html2pdf to render the HTML element to an image before converting it to PDF. The scale is set to 2, which means that the image is rendered at double the size of the original element. The logging property is set to true, which enables logging to the console. The dpi property is set to 192, which sets the resolution of the rendered image. The letterRendering property is set to true, which enables better rendering of letter shapes in the image.
The jsPDF configuration object is used to set options for the jsPDF library, which is used by html2pdf to generate the PDF file. The unit is set to millimeters (mm), the format is set to legal, and the orientation is set to portrait.
Finally, the pdf.print() method is called, which opens the generated PDF file in a new window for printing.