Create React-Reservation-form.js
This is a React component named Reservation that renders a form with two input fields: a checkbox for "isGoing" and a number input for "numberOfGuests". The component's state is initialized in the constructor with two properties: "isGoing" set to true and "numberOfGuests" set to 2.

The "handleInputChange" function is bound to "this" in the constructor and is called whenever the input fields change. It checks the type of the input field and updates the state accordingly using the "setState" method with the computed property name syntax to update the state property dynamically.

In the render method, the form is rendered with two label/input pairs. The first input is a checkbox with the "name" property set to "isGoing" and the "checked" property set to the "isGoing" state property. The "onChange" property is set to the "handleInputChange" method, which will be called when the checkbox is clicked.

The second input is a number input with the "name" property set to "numberOfGuests" and the "value" property set to the "numberOfGuests" state property. The "onChange" property is also set to the "handleInputChange" method, which will be called whenever the number input is changed.

Overall, this component allows users to specify if they are attending an event and how many guests they will be bringing, with the state of the form being stored and updated in the component's state.